+++ checkpoint/orbax/checkpoint/experimental/v1/_src/training/metadata/types.py
@@ -15,6 +15,7 @@
 """Metadata for `training.Checkpointer`."""
 
 import dataclasses
+from orbax.checkpoint._src.metadata import checkpoint_info
 from orbax.checkpoint.experimental.v1._src.metadata import types as metadata_types
 from orbax.checkpoint.experimental.v1._src.tree import types as tree_types
 
@@ -22,7 +23,9 @@ from orbax.checkpoint.experimental.v1._src.tree import types as tree_types
 CheckpointableMetadataT = metadata_types.CheckpointableMetadataT
 
 
-@dataclasses.dataclass(frozen=True, kw_only=True)
+CheckpointInfo = checkpoint_info.CheckpointInfo
+
+
 class CheckpointMetadata(
     metadata_types.CheckpointMetadata[CheckpointableMetadataT]
 ):
@@ -31,17 +34,46 @@ class CheckpointMetadata(
   Like its parent, the class has a `metadata` attribute that is a generic type.
 
   See superclass documentation for more information, and for a list of base
-  attributes.
+  attributes. This class defines several additional attributes that are relevant
+  to checkpoints in a sequence, but not necessarily to a singular checkpoint in
+  isolation.
 
-  Attributes:
+  Additional attributes:
+    step: The step number of the checkpoint.
     metrics: User-provided metrics for the step (e.g. loss, accuracy, etc.)
   """
-  metrics: tree_types.JsonType | None = None
+
+  def __init__(
+      self,
+      step: int,
+      *,
+      metadata: CheckpointableMetadataT,
+      init_timestamp_nsecs: int | None = None,
+      commit_timestamp_nsecs: int | None = None,
+      custom_metadata: tree_types.JsonType | None = None,
+      metrics: tree_types.JsonType | None = None,
+  ):
+    super().__init__(
+        metadata=metadata,
+        init_timestamp_nsecs=init_timestamp_nsecs,
+        commit_timestamp_nsecs=commit_timestamp_nsecs,
+        custom_metadata=custom_metadata,
+    )
+    self._step = step
+    self._metrics = metrics
+
+  @property
+  def step(self) -> int:
+    return self._step
+
+  @property
+  def metrics(self) -> tree_types.JsonType | None:
+    return self._metrics
 
 
 @dataclasses.dataclass(frozen=True, kw_only=True)
 class RootMetadata:
-  """Metadata of a checkpoint at root level (contains all steps).
+  """Metadata of a sequence of checkpoint at root level (contains all steps).
 
   Attributes:
     custom_metadata: User-provided custom metadata. An arbitrary